Silicon/AMD/Styx: update ArmTrustedFirmware.bin
The ARM Trusted Firmware build we have been using up until now was built
with optimizations disabled (which means every variable manipulation
involves a load, the operation itself and a store), and runs with the
MMU disabled, making it needlessly slow.

This appears to be due to the fact that
a) the page tables are not set up correctly, so not all memory can be
   accessed from EL3
b) the handling of SMC service calls does not take into account that
   these calls may be made with the MMU off (e.g., by UEFI PEI).

These issues have been fixed in the source code, which should hopefully
remove any performance bottlenecks that may become more noticeable now
that we are going to call into the secure firmware more often to perform
Spectre variant 2 mitigations, which have been backported as well.

So update the binary image to a RELEASE build that was created with
optimizations enabled, and has the above fixes incorporated.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>
Tested-by: Marc Zyngier <marc.zyngier@arm.com>