Platforms/RPi3: Restructure platform in preparation for Pi 4
Similar to what is being done in edk2-platforms, elements of the Pi 3
platform that can be factorized for use with the Pi 4 are factorized.

For non-OSI this only applies to the Logo driver, as the Device Tree
and the Trusted Firmware are of course too platform specific to be
factorized.

Signed-off-by: Pete Batard <pete@akeo.ie>
Reviewed-by: Philippe Mathieu-Daude <philmd@redhat.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>