Platform/RaspberryPi: Update TF-A to v2.5
This is a run-of-the-mill update of the TF-A binaries used by the
Raspberry Pi 3 and Raspberry Pi 4 platforms, based on the recently
released TF-A 2.5.

These binaries were built in an open and verifiable manner through
a GitHub Actions build script (https://github.com/pftf/pitf).

It should be noted that we are only updating the binaries due to the
existing ones getting a bit old, in case some of the ARM erratas and
changes, that have been included in the past two TF-A releases, may
benefit us. However, unless there are changes of direct interest for
the Pi platform, we are not planning to update these binaries for
each TF-A release.

Tested for regression on Pi 3 Model B and Pi 4 Model B.

Signed-off-by: Pete Batard <pete@akeo.ie>
Acked-by: Samer El-Haj-Mahmoud <Samer.El-Haj-Mahmoud@arm.com>
Acked-by: Ard Biesheuvel <ardb@kernel.org>