Emulator/X86EmulatorDxe: Replace with MultiArchUefiPkg build
Replace the old X86EmulatorDxe with one built from
https://github.com/intel/MultiArchUefiPkg. This is a much more modern,
recent implementation that's more reliable and is actively maintained.

Add driver binaries for both AArch64 and RISCV64, along with the
LoadOpRom application.

Update the LICENSE file with the GPL text: although MultiArchUefiPkg is
licensed under the LGPL, it uses Unicorn
(https://github.com/intel/unicorn-for-efi) which contains components
licensed under the GPL. As a result, the overall binary is licensed
under the GPL.

Signed-off-by: Rebecca Cran <rebecca@bsdio.com>